refactor: 예외 핸들러를 전역과 비즈니스 범위로 분리
